+++ src/test/kotlin/blackjack/domain/card/CardTest.kt
@@ -1,8 +1,5 @@
-package blackjack.domain
+package blackjack.domain.card
 
-import blackjack.domain.card.Card
-import blackjack.domain.card.CardNumber
-import blackjack.domain.card.CardPattern
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
