+++ src/test/kotlin/blackjack/domain/person/DealerTest.kt
@@ -1,11 +1,9 @@
-package blackjack.domain
+package blackjack.domain.person
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.CardNumber
 import blackjack.domain.card.CardPattern
 import blackjack.domain.card.Deck
-import blackjack.domain.person.Dealer
-import blackjack.domain.person.Hand
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
